refactor: test case for the manufacturing operartion
